Create TERADATA 2019
The Project for this course is based on the Teradata University 2019 national student competition. This competition is open to undergraduate students high schools, community colleges, and universities across America. Teradata partnered with Hire Heroes to develop a practical project for students. Competitors were asked to examine, analyze, and visualize the non-profit organization's data and business questions and present their findings and recommendations. The data and questions were used for this class project. This class project counts toward 40% of the grade for BAS 270, the capstone course for the Business Analytics AAS degree.